ttaylorr's requests
- Changed fields from underscore seperator to camelCase
- Fixed missing roles on CachedMessage
- Moved MessageComponent (previously MessagePart) out of IncomingMessageData
