+++ src/main/java/pro/beam/api/resource/chat/events/data/IncomingWidgetData.java
@@ -2,15 +2,16 @@ package pro.beam.api.resource.chat.events.data;
 
 import java.util.List;
 
+import com.google.gson.annotations.SerializedName;
 import pro.beam.api.resource.BeamUser;
 import pro.beam.api.resource.chat.AbstractChatEvent;
 
 public class IncomingWidgetData extends AbstractChatEvent.EventData {
     public int channel;
     public String id;
-    public String user_name;
-    public int user_id;
-    public List<BeamUser.Role> user_roles;
+    @SerializedName("user_name") public String userName;
+    @SerializedName("user_id") public int userId;
+    @SerializedName("user_roles") public List<BeamUser.Role> userRoles;
     public String message;
 
     public String getMessage() {
